requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
Django==3.2.6
django-bootstrap-form==3.4
django-crispy-forms==1.12.0
django-filter==2.4.0
django-heroku==0.3.1
django-phonenumber-field==5.2.0
django-widget-tweaks==1.4.8
gunicorn==20.1.0
phonenumbers==8.12.33
Pillow==8.3.1
psycopg2==2.9.1
python-decouple==3.5
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
whitenoise==5.3.0
